refactor: split up layout into layout and app chore: - Slight change to tsconfig - Removed some commented out code
