+++ src/components/Header/Header.jsx
@@ -7,177 +7,6 @@ import useWorkspaceStore from '../../store/useWorkspaceStore'
 import Title from '@/components/Title'
 import * as Blockly from 'blockly/core'
 
-// Example XMLs (Blockly)
-// const tutorialExamples = [
-//   {
-//     label: 'Example: draw a parametric plane in point normal form',
-//     xml: `
-// <xml xmlns="https://developers.google.com/blockly/xml">
-//   <block type="parametric_plane" x="24" y="24">
-//     <value name="point">
-//       <block type="linalg_vec3">
-//         <field name="X">-1</field>
-//         <field name="Y">3</field>
-//         <field name="Z">4</field>
-//       </block>
-//     </value>
-//     <value name="norm">
-//       <block type="linalg_vec3">
-//         <field name="X">1</field>
-//         <field name="Y">-1</field>
-//         <field name="Z">3</field>
-//       </block>
-//     </value>
-//   </block>
-// </xml>
-//
-//     `,
-//   },
-//   {
-//     label: 'Example: Vector equation line r = r0 + td form',
-//     xml:
-//       <xml xmlns="https://developers.google.com/blockly/xml">
-//         <block type="geo_vector" x="24" y="24">
-//           <value name="pos">
-//             <block type="linalg_vec3">
-//               <field name="X">5</field>
-//               <field name="Y">4.5</field>
-//               <field name="Z">1</field>
-//             </block>
-//           </value>
-//           <value name="dir">
-//             <block type="linalg_vec3">
-//               <field name="X">1</field>
-//               <field name="Y">2</field>
-//               <field name="Z">-2</field>
-//             </block>
-//           </value>
-//         </block>
-//       </xml>
-//         ,
-//   },
-//   {
-//     label: 'Example: spheres',
-//     xml: `
-//       <xml xmlns="https://developers.google.com/blockly/xml">
-//         <variables>
-//           <variable obj3D="true" id="s1">sphere1</variable>
-//           <variable obj3D="true" id="s2">sphere2</variable>
-//         </variables>
-//         <block type="variables_set_obj3D" x="24" y="24">
-//           <field name="VAR">sphere1</field>
-//           <value name="VALUE">
-//             <block type="geo_sphere">
-//               <value name="center">
-//                 <block type="linalg_vec3">
-//                   <field name="X">0</field>
-//                   <field name="Y">0</field>
-//                   <field name="Z">0</field>
-//                 </block>
-//               </value>
-//               <value name="radius">
-//                 <block type="scalar">
-//                   <field name="VAL">1.5</field>
-//                 </block>
-//               </value>
-//             </block>
-//           </value>
-//         </block>
-//         <block type="variables_set_obj3D" x="24" y="80">
-//           <field name="VAR">sphere2</field>
-//           <value name="VALUE">
-//             <block type="geo_sphere">
-//               <value name="center">
-//                 <block type="linalg_vec3">
-//                   <field name="X">3</field>
-//                   <field name="Y">2</field>
-//                   <field name="Z">1</field>
-//                 </block>
-//               </value>
-//               <value name="radius">
-//                 <block type="scalar">
-//                   <field name="VAL">0.8</field>
-//                 </block>
-//               </value>
-//             </block>
-//           </value>
-//         </block>
-//       </xml>
-//     `,
-//   },
-//   {
-//     label: 'Example: mixed objects',
-//     xml: `
-//       <xml xmlns="https://developers.google.com/blockly/xml">
-//         <variables>
-//           <variable obj3D="true" id="obj1">point</variable>
-//           <variable obj3D="true" id="obj2">sphere</variable>
-//           <variable obj3D="true" id="obj3">vector</variable>
-//         </variables>
-//         <block type="variables_set_obj3D" x="24" y="24">
-//           <field name="VAR">point</field>
-//           <value name="VALUE">
-//             <block type="geo_point">
-//               <value name="pos">
-//                 <block type="linalg_vec3">
-//                   <field name="X">1</field>
-//                   <field name="Y">1</field>
-//                   <field name="Z">1</field>
-//                 </block>
-//               </value>
-//             </block>
-//           </value>
-//         </block>
-//         <block type="variables_set_obj3D" x="24" y="80">
-//           <field name="VAR">sphere</field>
-//           <value name="VALUE">
-//             <block type="geo_sphere">
-//               <value name="center">
-//                 <block type="linalg_vec3">
-//                   <field name="X">-2</field>
-//                   <field name="Y">0</field>
-//                   <field name="Z">0</field>
-//                 </block>
-//               </value>
-//               <value name="radius">
-//                 <block type="scalar">
-//                   <field name="VAL">1.2</field>
-//                 </block>
-//               </value>
-//             </block>
-//           </value>
-//         </block>
-//         <block type="variables_set_obj3D" x="24" y="136">
-//           <field name="VAR">vector</field>
-//           <value name="VALUE">
-//             <block type="geo_vector">
-//               <value name="pos">
-//                 <block type="linalg_vec3">
-//                   <field name="X">0</field>
-//                   <field name="Y">0</field>
-//                   <field name="Z">0</field>
-//                 </block>
-//               </value>
-//               <value name="dir">
-//                 <block type="linalg_vec3">
-//                   <field name="X">2</field>
-//                   <field name="Y">-1</field>
-//                   <field name="Z">1</field>
-//                 </block>
-//               </value>
-//               <value name="scale">
-//                 <block type="scalar">
-//                   <field name="VAL">2.5</field>
-//                 </block>
-//               </value>
-//             </block>
-//           </value>
-//         </block>
-//       </xml>
-//     `,
-//   },
-// ]
-
 // Default guides (used if JSON not found)
 const fallbackGuides = [
   {
