--- src/pages/Home.jsx
@@ -1,17 +1,0 @@
-import { Button } from "../components/ui/button"
-
-const Home = () => {
-  return (
-    <>
-    <div>
-      This is Home
-    </div>
-
-    <div>
-      <Button>Click me</Button>
-    </div>
-    </>
-  )
-}
-
-export default Home
