+++ tsconfig.app.json
@@ -1,12 +1,12 @@
 {
   "compilerOptions": {
-    // ...
+    "composite": true,
     "baseUrl": ".",
     "paths": {
-      "@/*": [
-        "./src/*"
-      ]
-    }
-    // ...
-  }
+      "@/*": ["./src/*"]
+    },
+    "allowJs": true,
+    "checkJs": false
+  },
+  "include": ["src"]
 }
